First Prototype of Tenjin
This prototype so far is a moodle block, with an HTML link that directs the user (manager or teacher) of a course to a form that allows them to upload a '.csv' type of file. This first prototype was developed by the group member: Shahanaz Alex and would serve as a foundation for the next prototype.